luajit, acpitests: Fix build on hardfloat arm
luajit as well as acpitests build system has mind of its own and does not
respect CC and friends in makefiles so we have to inject
the CFLAGS via EXTRA_OEMAKE, some of ABI defining params
e.g. float-abi selection is mentioned in TUNE_CCARGS and
not in TOOLCHAIN_OPTIONS. This causes build to go for softfloat
build and that is not what we want.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>