Phoronix Test Suite package
This is the package recipe for the phoronix test suite v5.2.1. The
tool is for benchmarking and testing of the system.

Signed-off-by: Todd Brandt <todd.e.brandt@linux.intel.com>
----
 meta-oe/recipes-benchmark/phoronix-test-suite/phoronix-test-suite_5.2.1.bb | 24 ++++++++++++++++++++++++
 1 file changed, 24 insertions(+)

v2: changes based on Ross Burton's comments. However, I kept this line:
S = "${WORKDIR}/phoronix-test-suite", since the tarball folder doesn't
include the version number.
v3: changed the license file locatoin and moved the RDEPENDS field down
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>